Pin a Build by BuildConfigId and Build Number
Added a Function to Pin a Build by providing its BuildConfigId and
Buildnumber. A message string can be passed to the function, which will
show up as comment to the pinned build.

Revert "Pin a Build by ConfigId and BuildNumber"

This reverts commit
f8063d384b17c6d4dfb04d4be5abe7b9135cc837.

Pin a Build by ConfigId and BuildNumber

Added a Function to Pin a Build
by providing the ConfigId and
BuildNumber. You can provide a Message
which will show up as comment to
the pinned Build.